refactor: rename location model
