+++ src/domain/models/LocationsModel.ts
