+++ src/domain/models/index.ts
@@ -1,2 +1,2 @@
 export { default as AddAccountModel } from './addAccountModel';
-export { default as ListLocationsModel } from './listLocationsModel';
+export { default as LocationsModel } from './LocationsModel';
